ui: Keep dropdown positioning with popover
