+++ crates/ui/src/lib.rs
@@ -9,7 +9,6 @@ mod icon;
 mod index_path;
 #[cfg(any(feature = "inspector", debug_assertions))]
 mod inspector;
-mod popup_position;
 mod root;
 mod sizing;
 mod styled;
